helpers: fix imports and helpers.exit_code location
code expects to be able to assign to helpers.exit_code (which stopped
working correctly when exit_code was moved to helpers.misc module).
